+++ vdu_controls/vdu_controls/preset_controller.py
@@ -1,0 +1,59 @@
+from __future__ import annotations
+
+import glob
+import os
+from typing import Dict, List
+
+from vdu_controls.constants import CONFIG_DIR_PATH
+from vdu_controls.preset import Preset
+
+
+class PresetController:
+    def __init__(self) -> None:
+        self.presets: Dict[str, Preset] = {}
+
+    def reinitialize(self):
+        self.presets = {}
+
+    def find_presets_map(self) -> Dict[str, Preset]:
+        presets_still_present = []
+        # Use a stable order for the files - alphabetical filename.
+        for path_str in sorted(glob.glob(CONFIG_DIR_PATH.joinpath("Preset_*.conf").as_posix()), key=os.path.basename):
+            preset_name = os.path.splitext(os.path.basename(path_str))[0].replace('Preset_', '').replace('_', ' ')
+            if preset_name not in self.presets:
+                preset = Preset(preset_name)
+                preset.load()
+                self.presets[preset_name] = preset
+            presets_still_present.append(preset_name)
+        for preset_name in list(self.presets.keys()):
+            if preset_name not in presets_still_present:
+                del self.presets[preset_name]
+        # If Order_Presets.conf exists, reorder according to the CSV Preset names it holds.
+        order_presets_path = CONFIG_DIR_PATH.joinpath("Order_Presets.conf")
+        if order_presets_path.exists():
+            ordering = order_presets_path.read_text().split(",")
+            all_presets = list(self.presets.values())
+            # Use the Preset-name's position in the Order_Presets.conf CSV as the key-value (or zero if not in the CSV)
+            all_presets.sort(key=lambda obj: ordering.index(obj.name) if obj.name in ordering else 0)
+            self.presets = {}
+            for preset in all_presets:
+                self.presets[preset.name] = preset
+        return self.presets
+
+    def save_order(self, ordering: List[str]) -> None:
+        order_presets_path = CONFIG_DIR_PATH.joinpath("Order_Presets.conf")
+        order_presets_path.write_text(','.join(ordering))
+
+    def save_preset(self, preset: Preset) -> None:
+        preset.save()
+        self.presets[preset.name] = preset
+
+    def delete_preset(self, preset: Preset) -> None:
+        preset.delete()
+        del self.presets[preset.name]
+
+    def get_preset(self, preset_number: int) -> Preset | None:
+        presets = self.find_presets_map()
+        if preset_number < len(presets):
+            return list(presets.values())[preset_number]
+        return None
